+++ src/pebble/infrastructure/api/models/habit/__init__.py
@@ -1,0 +1,7 @@
+from .create_habit_request_model import CreateHabitRequest
+from .habit_response_model import HabitResponse
+
+__all__ = [
+    "CreateHabitRequest",
+    "HabitResponse",
+]